Split out worklet and worker tests
